pyproject.toml
[tool.poetry]
name = "bulletin-board-bot"
version = "0.1.0"
description = ""
authors = ["Kudyakov Artem <t3m8ch@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.18.0"
aiogram = "^2.13"
urlpath = "^1.1.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
ipython = "^7.24.1"

[tool.poetry.scripts]
bot = "bulletin_board_bot.main:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
